@@ -8,17 +8,23 @@ license = "MIT"
 
 [tool.poetry.dependencies]
 python = "^3.9"
-# 3rd party
-lark-parser = "^0.11.1"
 
-[tool.poetry.dev-dependencies]
-black = "^20.8b1"
-flake8 = "^3.8.3"
-ipython = "^7.31.1"
-mypy = "^0.800"
-pytest = "^6.2.0"
-pytest-cov = "^2.10.1"
+# 3rd party
+lark = "^1.1.7"
+
+[tool.poetry.group.dev.dependencies]
+black = "^23.7"
+flake8 = "^6.1.0"
+ipython = "^8.1.0"
+isort = "^5.12.0"
+mypy = "^1.5.0"
+pytest = "^7.4.0"
+pytest-cov = "^4.1.0"
 
 [build-system]
 build-backend = "poetry.core.masonry.api"
 requires = ["poetry-core>=1.0.0"]
+
+[tool.isort]
+multi_line_output = 3
+profile = "black"
