@@ -1,15 +1,14 @@
-# Church numbers
-# n(F) => call F n times
-ZERO = lambda f: lambda x: x
+"""
+Book of Changes:
+
+Therefore there is in the Changes the Great Primal Beginning.
+This generates the two primary forces.
+The two primary forces generate the four images.
+The four images generate the eight trigrams
+"""
+
+
+ZERO = lambda _: lambda x: x
 ONE = lambda f: lambda x: f(x)
-TWO = lambda f: lambda x: f(f(x))
-THREE = lambda f: lambda x: f(f(f(x)))
-FIVE = lambda f: lambda x: f(f(f(f(f(x)))))
-
-
-# conditions
-TRUE = lambda x: lambda y: x
-FALSE = lambda x: lambda y: y
-IF = lambda b: b  # reduced from `lambda b: lambda x: lambda y: b(x)(y)`
-NOT = lambda b: b(FALSE)(TRUE)
-IS_ZERO = lambda n: n(lambda _: FALSE)(TRUE)
+TRUE = lambda x: lambda _: x
+FALSE = lambda _: lambda y: y
